Quantize cropping for pixel-perfect texture placement
The texture crops had an error of up to 1/2 scaled pixel size
since the requested rectangle was not considering pixel
quantization. This was visible at high zoom levels or with
maps that have a large scaling of pixels.

Now the requested rectangle is a multiple of the scaled pixel
size.