Fix loss of custom tints when deleting active_tint_selection map
When the map that was selected in the UI got deleted, the selection
fell back to None. This meant that a single "All" setting got
selected, which made the other remaining maps lose their custom tint.

Now the active selection switches to one of the other maps to prevent
accidentally losing the settings (if there are other maps left).