Try to handle egui color changes for old serialized session files
Without any conversion, an old saved value appears as a completely
different color because it's assumed to be a gamma-premultiplied
value (which it wasn't when serialized with egui < 0.32).
This then lead e.g. the color picker to do wrong conversions and
mess up the tint color.

We now check if the session was written with the latest or an old
version and fix the misinterpreted color values if needed.

Note that while the values are correct after conversion, the image
can still appear darker due to the blending changes in egui.

Relates to: https://github.com/emilk/egui/pull/5824
Relates to: https://github.com/emilk/egui/pull/7311